Happy New Year from Giperion! Here's update for you P.S: Please check CWeapon::UpdateXForm : 131 - 153 lines, it's broken * ADD: (Editor) Pre-build script for xrPostprocessEditor * ADD: (Editor) Readme.txt for xrPostprocessEditor * ADD: (Editor) xrCore project as reference * EDIT: (Editor) Build rules cleanup, debug configuration supported * FIX: (Editor) We have way, to get rid of errors when opening forms, but not a perfect solution (switch managed projects to AnyCPU configurations and rebuild) * FIX: (Editor) BasicPostProcessAnimator::Load is changed to most stable version (comment test code) * FIX: xrQIK can be compiled in Debug configuration * FIX: xrEngine, xrScript and xrGame can be compiled in Debug configuration, but not for long, plz hurry! * FIX: GameState packet import/export is now correctly synchronized * FIX: CEnvironment::mods_load: Function things, that size_t == 4, but IRL (for x86_64) it's a 8. Corrected.

+++ code/engine.vc2008/xrEngine/Environment_misc.cpp
@@ -452,11 +452,11 @@ void	CEnvironment::mods_load()
 		IReader*	fs = FS.r_open(path);
 		u32			id = 0;
 		u32 ver = 0x0015;
-		size_t sz;
+		u32 sz;
 
 		while (0 != (sz = fs->find_chunk(id)))
 		{
-			if (id == 0 && sz == sizeof(size_t))
+			if (id == 0 && sz == sizeof(u32))
 			{
 				ver = fs->r_u32();
 			}
